backend/requirements.txt

flask
flask-cors
transformers
torch
numpy
scikit-learn
shap
lime
pymongo
newspaper3k
beautifulsoup4
python-dotenv
requests
pytest
